+++ CHANGELOG.md
@@ -27,6 +27,8 @@
     - Files inherit from `OTOS::iostream`.
 - New Examples which demonstrate how to use the OTOS.
 - Adds *iostreams* to `Kernel`.
+- *iostream* has overloads for adding:
+    - Null terminated string.
 
 ### Fixed Issues:
 - #12: Update core_stm32 for STM32L0 Devices